+++ src/test/index.ts
@@ -1,0 +1,2 @@
+import "./store/dao.spec";
+import "./store/reusable-dao.spec";